Move tests to pytest
